Fix occasional unwanted display rotations.
A number of people were complaining about display rotations when holding
the phone perfectly upwards and speaker activity, such as a ringtone
playing or a YouTube video playing. Fix that by

- Adding only one accelerometer to the sensor list.
Previously, both the AK8973 and the KXTF9 accelerometer were advertised
to the sensor service. Only the KXTF9 was actually used, but the AK8973
one was listed first, so the service picked up this one. This led to a
mismatch of sensor values and advertised value range.

- Dropping the KXTF9 orientation sensor code.
Due to limited precision and the fact that the KXTF9 orientation sensor
is 2-axis only, we weren't using the sensor anyway (at least we didn't
advertise it). Nevertheless, the kernel driver sent events, which we
handled and pushed to the service. While doing so, we shared the event
structure, which (at least could) lead to wrong data being sent to the
service (e.g. accelerometer precision was always sent as UNRELIABLE =
0). Stop doing that and handle only the accelerometer in the KXTF9
driver.

Change-Id: I3b3f1955c855ab1a81785767ce616d33713de341

Conflicts:

	libsensors/SensorKXTF9.cpp

Conflicts:

	libsensors/SensorKXTF9.cpp